Create node constructors
