+++ onnxscript/ir/__init__.py
@@ -70,8 +70,9 @@ __all__ = [
     # Conversion functions
     "from_proto",
     "to_proto",
-    # IR Tensor initializer
+    # Convenience constructors
     "tensor",
+    "node",
     # Pass infrastructure
     "passes",
     # IO
@@ -80,7 +81,7 @@ __all__ = [
 ]
 
 from onnxscript.ir import convenience, external_data, passes, serde, traversal
-from onnxscript.ir._convenience import tensor
+from onnxscript.ir._convenience._constructors import node, tensor
 from onnxscript.ir._core import (
     Attr,
     AttrFloat32,
