+++ onnxscript/ir/_convenience/__init__.py
@@ -14,16 +14,11 @@ __all__ = [
     "replace_all_uses_with",
 ]
 
-import typing
 from typing import Mapping, Sequence, Union
 
-import numpy as np
 import onnx
 
-from onnxscript.ir import _core, _enums, _protocols, serde, tensor_adapters
-
-if typing.TYPE_CHECKING:
-    import numpy.typing as npt
+from onnxscript.ir import _core, _enums, _protocols, serde
 
 SupportedAttrTypes = Union[
     str,
@@ -289,94 +284,6 @@ def replace_all_uses_with(
     for value, replacement in zip(values, replacements):
         for user_node, index in tuple(value.uses()):
             user_node.replace_input_with(index, replacement)
-
-
-def tensor(
-    value: npt.ArrayLike
-    | onnx.TensorProto
-    | _protocols.DLPackCompatible
-    | _protocols.ArrayCompatible,
-    dtype: _enums.DataType | None = None,
-    name: str | None = None,
-    doc_string: str | None = None,
-) -> _protocols.TensorProtocol:
-    """Create a tensor value from an ArrayLike object or a TensorProto.
-
-    The dtype must match the value. Reinterpretation of the value is
-    not supported, unless if the value is a plain Python object, in which case
-    it is converted to a numpy array with the given dtype.
-
-    :param:`value` can be a numpy array, a plain Python object, or a TensorProto.
-
-    Example::
-
-        >>> from onnxscript import ir
-        >>> import numpy as np
-        >>> import ml_dtypes
-        >>> import onnx
-        >>> ir.tensor(np.array([1, 2, 3], dtype=np.int16))
-        Tensor<INT16,[3]>(array([1, 2, 3], dtype=int16), name=None)
-        >>> ir.tensor([1, 2, 3], dtype=ir.DataType.BFLOAT16)
-        Tensor<BFLOAT16,[3]>(array([1, 2, 3], dtype=bfloat16), name=None)
-        >>> tp_tensor = ir.tensor(onnx.helper.make_tensor("tensor", onnx.TensorProto.FLOAT, dims=[], vals=[0.5]))
-        >>> tp_tensor.numpy()
-        array(0.5, dtype=float32)
-        >>> import torch
-        >>> ir.tensor(torch.tensor([1.0, 2.0]), name="torch_tensor")
-        TorchTensor<FLOAT,[2]>(tensor([1., 2.]), name='torch_tensor')
-
-    Args:
-        value: The numpy array to create the tensor from.
-        dtype: The data type of the tensor.
-        name: The name of the tensor.
-        doc_string: The documentation string of the tensor.
-
-    Returns:
-        A tensor value.
-
-    Raises:
-        ValueError: If the dtype does not match the value when value is not a plain Python
-            object like ``list[int]``.
-    """
-    if isinstance(value, _protocols.TensorProtocol):
-        if dtype is not None and dtype != value.dtype:
-            raise ValueError(
-                f"The dtype must match the value when value is a Tensor. dtype={dtype}, value.dtype={value.dtype}. "
-                "You do not have to specify the dtype when value is a Tensor."
-            )
-        return value
-    if isinstance(value, onnx.TensorProto):
-        tensor_ = serde.deserialize_tensor(value)
-        if name is not None:
-            tensor_.name = name
-        if doc_string is not None:
-            tensor_.doc_string = doc_string
-        if dtype is not None and dtype != tensor_.dtype:
-            raise ValueError(
-                f"The dtype must match the value when value is a TensorProto. dtype={dtype}, value.data_type={tensor_.dtype}"
-                "You do not have to specify the dtype when value is a TensorProto."
-            )
-        return tensor_
-    elif str(type(value)) == "<class 'torch.Tensor'>":
-        # NOTE: We use str(type(...)) and do not import torch for type checking
-        # as it creates overhead during import
-        return tensor_adapters.TorchTensor(value, name=name, doc_string=doc_string)  # type: ignore[arg-type]
-    elif isinstance(value, (_protocols.DLPackCompatible, _protocols.ArrayCompatible)):
-        return _core.Tensor(value, dtype=dtype, name=name, doc_string=name)
-
-    # Plain Python object
-    if dtype is not None:
-        numpy_dtype = dtype.numpy()
-    else:
-        numpy_dtype = None
-    array = np.array(value, dtype=numpy_dtype)
-    return _core.Tensor(
-        array,
-        dtype=dtype,
-        shape=_core.Shape(array.shape),
-        name=name,
-        doc_string=name,
-    )
 
 
 def create_value_mapping(graph: _core.Graph) -> dict[str, _core.Value]:
